+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Add drop-down filters for constellation and rarity to extractions (#12)
+
 ### Changed
 
 ### Fixed